refactored network.game into more files
