+++ src/engine/config.game.ts
@@ -2,7 +2,7 @@ import { defineConfig } from "./module/module.common";
 import { GameAudioModule, GameAudioSystem } from "./audio/audio.game";
 import { ApplyInputSystem, InputModule, ResetInputSystem } from "./input/input.game";
 import { PhysicsModule, PhysicsSystem } from "./physics/physics.game";
-import { InboundNetworkSystem, NetworkModule, OutboundNetworkSystem } from "./network/network.game";
+import { NetworkModule } from "./network/network.game";
 import { ActionMappingSystem } from "./input/ActionMappingSystem";
 import { FirstPersonCameraModule, FirstPersonCameraSystem } from "../plugins/FirstPersonCamera";
 import { PhysicsCharacterControllerModule, PlayerControllerSystem } from "../plugins/PhysicsCharacterController";
@@ -27,6 +27,8 @@ import { AnimationSystem } from "./animation/animation.game";
 import { NameSystem } from "./component/Name";
 import { InteractionModule, InteractionSystem } from "../plugins/interaction/interaction.game";
 import { NametagModule, NametagSystem } from "../plugins/nametags/nametags.game";
+import { InboundNetworkSystem } from "./network/inbound.game";
+import { OutboundNetworkSystem } from "./network/outbound.game";
 
 export default defineConfig<GameState>({
   modules: [
