+++ src/engine/network/NetworkAction.ts
@@ -1,0 +1,14 @@
+export enum NetworkAction {
+  Create,
+  Delete,
+  UpdateChanged,
+  UpdateSnapshot,
+  FullChanged,
+  FullSnapshot,
+  Prefab,
+  AssignPeerIdIndex,
+  InformPlayerNetworkId,
+  NewPeerSnapshot,
+  RemoveOwnershipMessage,
+  Command,
+}
